+++ example_runner.py
@@ -51,7 +51,7 @@ import example.tower.round_block_stucco_generator
 
 
 #------------------------------
-run_base = True
+run_base = False
 
 if run_base:
     print('Tower Examples')
@@ -88,7 +88,7 @@ if run_base:
     print('end tower_top')
 
 #------------------------------
-run_greebled = True
+run_greebled = False
 
 if run_greebled:
     print('start tower_greebled_example')
@@ -127,7 +127,7 @@ if run_greebled:
 
 
 #-----------------------------------
-run_alt = True
+run_alt = False
 
 if run_alt:
     print('start tower_alt_greebled_example')
@@ -139,7 +139,7 @@ if run_alt:
     print('end tower_alt_greebled_example')
 
 #-----------------------------------
-run_stucco = True
+run_stucco = False
 
 if run_stucco:
     print('start tower_stucco_example')